backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "A FastAPI backend for processing API data and passing it to the frontend."
authors = ["Ryan Partridge <rpartridge101@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
run-dev = "app.start:development"
run-prod = "app.start:production"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
python-dotenv = "^1.0.1"
appwrite = "^5.0.3"
pydantic-settings = "^2.3.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
hypothesis = "^6.105.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov-report term-missing --cov=app tests/"
